Move DriverFunction from h2gis-function to h2gis-api.
